+++ src/main/java/com/publicuhc/pluginframework/routing/RoutingModule.java
@@ -1,6 +1,8 @@
 package com.publicuhc.pluginframework.routing;
 
 import com.google.inject.AbstractModule;
+import com.publicuhc.pluginframework.routing.parser.DefaultRoutingMethodParser;
+import com.publicuhc.pluginframework.routing.parser.RoutingMethodParser;
 
 public class RoutingModule extends AbstractModule
 {
