+++ src/main/java/com/publicuhc/pluginframework/routing/parser/CommandOptionsParser.java
@@ -1,4 +1,4 @@
-package com.publicuhc.pluginframework.routing;
+package com.publicuhc.pluginframework.routing.parser;
 
 import org.bukkit.craftbukkit.libs.joptsimple.OptionParser;
 import org.bukkit.craftbukkit.libs.joptsimple.OptionSet;
