+++ src/test/java/com/publicuhc/pluginframework/routing/parser/DefaultCommandOptionsParserTest.java
@@ -1,6 +1,7 @@
-package com.publicuhc.pluginframework.routing;
+package com.publicuhc.pluginframework.routing.parser;
 
 import com.publicuhc.pluginframework.routing.exception.OptionMissingException;
+import com.publicuhc.pluginframework.routing.parser.DefaultCommandOptionsParser;
 import org.bukkit.craftbukkit.libs.joptsimple.OptionParser;
 import org.bukkit.craftbukkit.libs.joptsimple.OptionSet;
 import org.junit.Before;
